Реализована функциональность kanban board

+++ src/main/java/ru/miniprog/minicrm/repository/UserRepository.java
@@ -7,7 +7,7 @@ import ru.miniprog.minicrm.model.UserCrm;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserCrm, Integer> {
+public interface UserRepository extends JpaRepository<UserCrm, Long> {
     Optional<UserCrm> findByUsername(String username);
     boolean existsByUsername(String username);
     boolean existsByEmail(String email);
